Add files via upload
Tiny ImageNet is image classification problem.  In which the dataset contains 200 different objects and each object has 500 images for training purpose, In this model, the baseline model is developed which initially built for only 10 classes can be extended up to 200.  multiple convolutions and fully connected  layers are used in order to achieve 56% testing, 85%  validation and 93% training accuracy.